+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "FhirToDataLake/src/Microsoft.Health.Fhir.Synapse.Parquet.Native/vcpkg"]
-	path = FhirToDataLake/src/Microsoft.Health.Fhir.Synapse.Parquet.Native/vcpkg
+[submodule "FhirToDataLake/native/parquet/cpp/vcpkg"]
+	path = FhirToDataLake/native/parquet/cpp/vcpkg
 	url = ../vcpkg
