+++ FhirToDataLake/Microsoft.Health.Fhir.Synapse.sln
@@ -34,8 +34,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Health.Fhir.Synap
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Health.Fhir.Synapse.E2ETests", "test\Microsoft.Health.Fhir.Synapse.E2ETests\Microsoft.Health.Fhir.Synapse.E2ETests.csproj", "{01C19DA0-63E6-43E2-B421-12CC8BD7EB94}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Health.Fhir.Synapse.Parquet.UnitTests", "test\Microsoft.Health.Fhir.Synapse.Parquet.UnitTests\Microsoft.Health.Fhir.Synapse.Parquet.UnitTests.csproj", "{35755AAC-C867-499E-8F49-A141ADDD54FA}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Health.Fhir.Synapse.Common.UnitTests", "test\Microsoft.Health.Fhir.Synapse.Common.UnitTests\Microsoft.Health.Fhir.Synapse.Common.UnitTests.csproj", "{A46BBBEA-22C5-413C-81BF-5C29770A69EF}"
 EndProject
 Global
