+++ FhirToDataLake/azurepipelines/build-windows-native-libs.yml
@@ -7,23 +7,23 @@ steps:
 - task: CMake@1
   displayName: CMake generate configuration
   inputs:
-    workingDirectory: FhirToDataLake/src/Microsoft.Health.Fhir.Synapse.Parquet.Native
+    workingDirectory: FhirToDataLake/native/parquet/cpp
     cmakeArgs: " -B build -S . -DCMAKE_BUILD_TYPE=Release"
 
 - task: CMake@1
   displayName: CMake build
   inputs:
-    workingDirectory: FhirToDataLake/src/Microsoft.Health.Fhir.Synapse.Parquet.Native
+    workingDirectory: FhirToDataLake/native/parquet/cpp
     cmakeArgs: " --build build --config Release"
 
 - script: ctest -V
-  workingDirectory: FhirToDataLake/src/Microsoft.Health.Fhir.Synapse.Parquet.Native/build
+  workingDirectory: FhirToDataLake/native/parquet/cpp/build
   displayName: Run native tests
 
 # Pack target shared library and all dependencies to artifact.
 - script: | 
     xcopy /S /Q /F /Y build\src\Release $(System.DefaultWorkingDirectory)\native-windows-x64\
-  workingDirectory: FhirToDataLake/src/Microsoft.Health.Fhir.Synapse.Parquet.Native
+  workingDirectory: FhirToDataLake/native/parquet/cpp
   displayName: Copy native libraries
 
 - publish: $(System.DefaultWorkingDirectory)/native-windows-x64
