+++ FhirToDataLake/native/parquet/cpp/src/SchemaManager.h
