+++ FhirToDataLake/native/parquet/cpp/test/data/Expected/expected_patient.parquet
