+++ FhirToDataLake/native/parquet/cpp/test/data/Patient.ndjson
