+++ FhirToDataLake/native/parquet/cpp/test/data/patient_example_schema.json
