+++ FhirToDataLake/native/parquet/csharp/src/BuildParquet.nuspec
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <package xmlns="http://schemas.microsoft.com/packaging/2010/07/nuspec.xsd">
     <metadata>
-        <id>Microsoft.Health.Fhir.Synapse.Parquet</id>
+        <id>Microsoft.Health.Parquet</id>
         <version>1.0.0.0</version>
         <description>Parquet library built with Arrow cpp.</description>
 		<authors>Microsoft</authors>
