+++ FhirToDataLake/native/parquet/csharp/test/Microsoft.Health.Parquet.UnitTests.csproj
@@ -17,7 +17,7 @@
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
       <PrivateAssets>all</PrivateAssets>
     </PackageReference>
-    <PackageReference Include="Microsoft.Health.Fhir.Synapse.Parquet" Version="$(ParquetNativeLibVersion)" />
+    <PackageReference Include="Microsoft.Health.Parquet" Version="1.0.0.41" />
   </ItemGroup>
   <ItemGroup>
       <None Update="TestData\**\*.*">
