+++ FhirToDataLake/native/parquet/csharp/test/TestData/Patient.ndjson
