+++ FhirToDataLake/native/parquet/csharp/test/TestData/patient_example_schema.json
